+++ common/src/main/java/net/witcher_rpg/config/TweaksConfig.java
@@ -2,6 +2,7 @@ package net.witcher_rpg.config;
 
 public class TweaksConfig {
     public boolean ignore_items_required_mods = false;
-    public int adrenaline_max_amplifier = 10;
+    public int adrenaline_max_amplifier = 20;
     public int adrenaline_max_seconds_duration = 180;
+    public float battle_trance_damage_per_adrenaline_level = 0.05F;
 }